requirements.txt
Pillow 
hashlib
tkinter
tk